fix: presentations not displaying properly
