+++ api.Dockerfile
@@ -24,9 +24,11 @@ RUN apk add --no-cache \
     harfbuzz \
     ca-certificates \
     ttf-freefont \
+    font-noto \
     xvfb \
     libc6-compat \
     ffmpeg \
+    libreoffice \
     && rm -rf /var/cache/apk/*
 
 ENV PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true \
