+++ apps/api/src/file/file.constants.ts
@@ -2,6 +2,7 @@ export const REDIS_TTL = 59 * 60 * 1000;
 export const MAX_FILE_SIZE = 1024 * 1024 * 1024;
 export const MAX_VIDEO_SIZE = 5 * 1024 * 1024 * 1024;
 export const MAX_COURSE_TRAILER_VIDEO_SIZE = 200 * 1024 * 1024;
+export const PRESENTATION_PDF_PREVIEW_CONTENT_TYPE = "application/pdf";
 
 export const ALLOWED_MIME_TYPES = [
   "image/jpeg",
