+++ apps/api/src/news/news.controller.ts
@@ -36,6 +36,7 @@ import { RequirePermission } from "src/common/decorators/require-permission.deco
 import { CurrentUser } from "src/common/decorators/user.decorator";
 import { CurrentUserType } from "src/common/types/current-user.type";
 import { supportedLanguagesSchema } from "src/courses/schemas/course.schema";
+import { filePreviewQuerySchema, FilePreviewQuery } from "src/file/types/file-preview.type";
 import { getBaseFileTypePipe } from "src/file/utils/baseFileTypePipe";
 import { buildFileTypeRegex } from "src/file/utils/fileTypeRegex";
 import { ValidateMultipartPipe } from "src/utils/pipes/validateMultipartPipe";
@@ -104,15 +105,19 @@ export class NewsController {
   @Public()
   @Get("news-resource/:resourceId")
   @Validate({
-    request: [{ type: "param", schema: UUIDSchema, name: "resourceId" }],
+    request: [
+      { type: "param", schema: UUIDSchema, name: "resourceId" },
+      { type: "query", name: "preview", schema: filePreviewQuerySchema },
+    ],
   })
   async getNewsResource(
     @Param("resourceId") resourceId: UUIDType,
+    @Query("preview") preview: FilePreviewQuery,
     @CurrentUser() currentUser: CurrentUserType | undefined,
     @Req() req: Request,
     @Res() res: Response,
   ) {
-    return this.newsService.getNewsResource(req, res, resourceId, currentUser);
+    return this.newsService.getNewsResource(req, res, resourceId, currentUser, preview);
   }
 
   @Public()
