+++ apps/api/src/swagger/api-schema.json
@@ -6149,6 +6149,15 @@
               "format": "uuid",
               "type": "string"
             }
+          },
+          {
+            "name": "preview",
+            "required": false,
+            "in": "query",
+            "schema": {
+              "const": "pdf",
+              "type": "string"
+            }
           }
         ],
         "responses": {
@@ -6168,6 +6177,15 @@
             "in": "path",
             "schema": {
               "format": "uuid",
+              "type": "string"
+            }
+          },
+          {
+            "name": "preview",
+            "required": false,
+            "in": "query",
+            "schema": {
+              "const": "pdf",
               "type": "string"
             }
           }
@@ -11750,6 +11768,15 @@
               "format": "uuid",
               "type": "string"
             }
+          },
+          {
+            "name": "preview",
+            "required": false,
+            "in": "query",
+            "schema": {
+              "const": "pdf",
+              "type": "string"
+            }
           }
         ],
         "responses": {
@@ -12463,6 +12490,15 @@
             "in": "path",
             "schema": {
               "format": "uuid",
+              "type": "string"
+            }
+          },
+          {
+            "name": "preview",
+            "required": false,
+            "in": "query",
+            "schema": {
+              "const": "pdf",
               "type": "string"
             }
           }
