+++ apps/web/app/components/RichText/extensions/pdfPreview/constants.ts
@@ -1,0 +1,9 @@
+export const PDF_PREVIEW_NODE_TYPE = "pdf-preview";
+
+export const PDF_VIEWER_ACTION = {
+  RESET: "reset",
+  DOCUMENT_LOADED: "documentLoaded",
+  REQUEST_PAGE_CHANGE: "requestPageChange",
+  PRELOAD_SUCCESS: "preloadSuccess",
+  PRELOAD_ERROR: "preloadError",
+} as const;
