+++ apps/web/app/locales/lt/translation.json
@@ -2945,6 +2945,7 @@
       }
     },
     "presentation": {
+      "title": "Pristatymas",
       "ariaLabel": {
         "drag": "Vilkite pristatymo įdėjimą",
         "remove": "Pašalinti pristatymo įterpimą"
